Fix navigation to Service Explorer from the `generic_object` view
This issue is related to the newly converted Generic Object Definition
view to Treeview
`x_node` remembers the earlier node selected from the Generic Object
Definition tree and tries to map that node to the Services Tree.
Our best bet is to check params[:id] for a possible id value select the
tree node based on that.

Also fixed the quad icon url value for Services